+++ src/Resources/contao/dca/tl_ls_api_user.php
@@ -2,6 +2,7 @@
 
 namespace LeadingSystems\Api;
 
+use Contao\Backend;
 use Contao\DataContainer;
 use Contao\DC_Table;
 
@@ -103,6 +104,7 @@ $GLOBALS['TL_DCA']['tl_ls_api_user'] = array
             'eval'                    => array('mandatory' => true, 'rgxp' => 'extnd', 'nospace' => true, 'unique' => true, 'maxlength' => 64, 'tl_class'=>'w50'),
             'sql'                     => "varchar(64) BINARY NULL"
         ),
+        // @toDo Add Regular expression to avoid specific characters
         'password' => array
         (
             'label'                   => &$GLOBALS['TL_LANG']['tl_ls_api_user']['password'],
@@ -121,10 +123,10 @@ $GLOBALS['TL_DCA']['tl_ls_api_user'] = array
     )
 );
 
-class tl_ls_api_user extends \Backend {
+class tl_ls_api_user extends Backend {
     public function __construct() {
         parent::__construct();
-        $this->import('BackendUser', 'User');
+        $this->import('Contao\BackendUser', 'User');
     }
 
     /**