+++ CHANGELOG.md
@@ -1,7 +1,12 @@
-Leading Systems API bundle for Contao 4 changelog
+Leading Systems API bundle for Contao changelog
 ===========================================
 
-### 2.1.1 (nry: 2023-11-10)
+### 2.1.2 (2025-01-24)
+ * Removed TL_MODE
+ * Fix disable Referer-Check
+ * Fix undefined classes for Contao 5
+
+### 2.1.1 (2023-12-11)
  * Fix some issues for Contao 5
 
 ### 2.1.0 (2023-09-28)